mainMode.cppをはやした・make runSWDを追加した

+++ 2024/STMDev/F446RE_cubeide/src/unit/MotorDriver.hpp
@@ -14,7 +14,7 @@ class MotorDriver {
 
     void init();
 
-    void setVelocityFF(int16_t velX, int16_t velY, float velAng);
+    void setVelocityFF(int16_t velX, int16_t velY, int16_t velAng);
 
     void setMotors(int16_t M0, int16_t M1, int16_t M2, int16_t M3);
     // void setVelocity(int16_t velX, int16_t velY, float velAng);
